requirements.txt
openai
azure-identity
python-dotenv
azure-functions
azure-monitor-opentelemetry